Add files via upload
Plant orders and families based on NCBI taxonomy. The families that are present in the reference library are reported in black, those lacking in grey. The orders Chlorocystidales, Chlorosarcinales, Ignatiales, Oedogoniales, Oltmansiellopsidales, Microthamniales and Pyramimonadales are not reported as they do not contain any family. *Trebouxiaceae is not present in the reference databases as the corresponding accessions are recorded with their former order name "Microthamniales".